requirements.txt
requests
selenium
pyvirtualdisplay
BeautifulSoup4
antigate
pillow